likewise split test_powerups.py into plotly, matplotlib and both
